feat: update rank

+++ BusinessObjects/Models/Rank.cs
@@ -11,7 +11,7 @@ public partial class Rank
 
     public string? Description { get; set; }
 
-    public double? Discount { get; set; }
+    public decimal? Discount { get; set; }
 
     public virtual ICollection<User> Users { get; set; } = new List<User>();
 }
